requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.3.0
django-jsonify==0.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-access-policy==0.6.2
idna==2.10
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
